+++ examples/linux.rs
@@ -1,7 +1,3 @@
-extern crate embedded_hal;
-extern crate linux_embedded_hal;
-extern crate mcp49xx;
-
 use linux_embedded_hal::{Pin, Spidev};
 use mcp49xx::{Command, Mcp49xx};