+++ src/buffering.rs
@@ -1,18 +1,18 @@
-use {marker, private, Error};
+use crate::{marker, private, Error};
 
 #[doc(hidden)]
-pub trait BufferingSupport<E>: private::Sealed {
-    fn check_buffering_is_appropriate(buffered: bool) -> Result<(), Error<E>>;
+pub trait BufferingSupport<CommE, PinE>: private::Sealed {
+    fn check_buffering_is_appropriate(buffered: bool) -> Result<(), Error<CommE, PinE>>;
 }
 
-impl<E> BufferingSupport<E> for marker::Buffered {
-    fn check_buffering_is_appropriate(_buffered: bool) -> Result<(), Error<E>> {
+impl<CommE, PinE> BufferingSupport<CommE, PinE> for marker::Buffered {
+    fn check_buffering_is_appropriate(_buffered: bool) -> Result<(), Error<CommE, PinE>> {
         Ok(())
     }
 }
 
-impl<E> BufferingSupport<E> for marker::Unbuffered {
-    fn check_buffering_is_appropriate(buffered: bool) -> Result<(), Error<E>> {
+impl<CommE, PinE> BufferingSupport<CommE, PinE> for marker::Unbuffered {
+    fn check_buffering_is_appropriate(buffered: bool) -> Result<(), Error<CommE, PinE>> {
         if buffered {
             Err(Error::BufferingNotSupported)
         } else {